Move word table on course details page to its own component for reusability
