+++ ClientApp/src/app/courses/course-details/course-details.component.ts
@@ -1,28 +1,18 @@
-import { AfterViewInit, ViewChild } from '@angular/core';
 import { Component, OnInit } from '@angular/core';
-import { MatPaginator, PageEvent } from '@angular/material/paginator';
-import { MatTableDataSource } from '@angular/material/table';
 import { ActivatedRoute, Router } from '@angular/router';
 import { CourseService } from 'src/app/services/course.service';
 import { Course } from '../../models/course.model';
-import { Word } from '../../models/word.model';
 
 @Component({
   selector: 'app-course-details',
   templateUrl: './course-details.component.html',
   styleUrls: ['./course-details.component.css']
 })
-export class CourseDetailsComponent implements OnInit, AfterViewInit {
-  @ViewChild(MatPaginator) public paginator: MatPaginator;
-  public dataSource: MatTableDataSource<Word>;
-  public displayedColumns: string[] = ['Japanese', 'English', 'Sentence'];
+export class CourseDetailsComponent implements OnInit {
   public course: Course;
   public subscribed: string;
-  public selectedWords: Word[];
-  public selectedTableAction: string;
   public editMode: string;
   public manageContentText: string;
-  private DATA_CHUNK_SIZE = 500;
 
   public constructor(
   private route: ActivatedRoute, 
@@ -39,69 +29,15 @@ export class CourseDetailsComponent implements OnInit, AfterViewInit {
         this.subscribed = subscribedCourses.find(element => element.id === this.course.id) ? 'true' : 'false';
       });
     }
-    this.dataSource = new MatTableDataSource<Word>(this.course.words.slice(0, this.DATA_CHUNK_SIZE));
-    this.selectedWords = [];
     this.route.queryParams.subscribe((params) => {
       this.editMode = params.editMode;
       this.manageContentText = this.editMode === 'true' ? 'Display course content' : 'Edit course content';
     });
   }
 
-  public ngAfterViewInit() {
-    this.dataSource.paginator = this.paginator;
-  }
-
-  public checkForAdditionalTableData(pageEvent: PageEvent) {
-    if (!this.paginator.hasNextPage()) {
-      // Check if we have reached the end of our word list, expand by the data chunk size if not
-      // or expand until the end of the word list if so.
-      if (pageEvent.length + this.DATA_CHUNK_SIZE < this.course.words.length) {
-        this.dataSource = new MatTableDataSource<Word>(this.course.words.slice(0, pageEvent.length + this.DATA_CHUNK_SIZE));
-      } else {
-        this.dataSource = new MatTableDataSource<Word>(this.course.words);
-      }
-      this.dataSource.paginator = this.paginator;
-    }
-  }
-
-  public toggleWordStatus(word: Word) {
-    const wordIndex = this.selectedWords.findIndex(element => element.id === word.id);
-    if (wordIndex === -1) {
-      this.selectedWords.push(word);
-    } else {
-      this.selectedWords.splice(wordIndex, 1);
-    }
-  }
-
-  public changeSelectedTableAction(event: any) {
-    this.selectedTableAction = event.target.value;
-  }
-
-  public executeTableAction() {
-    // Placeholder switch statement to be implemented after
-    // database work for per user course management is done
-    switch (this.selectedTableAction) {
-      case null:
-        break;
-      case 'suspend':
-          console.log('suspend');
-          break;
-      case 'restore':
-        console.log('restore');
-        break;
-      case 'edit':
-        console.log('edit');
-        break;
-      case 'reset':
-        console.log('reset');
-        break;
-    }
-  }
-
   public updateCourse(): void {
     this.courseService.getCourse(this.course.id).subscribe((updatedCourse) => {
       this.course = updatedCourse;
-      this.dataSource = new MatTableDataSource<Word>(this.course.words.slice(0, this.DATA_CHUNK_SIZE));
     });
   }
