+++ ClientApp/src/app/courses/word-table/word-table.component.html
@@ -1,1 +1,48 @@
-<p>word-table works!</p>
+
+<div class="mat-elevation-z8 table-control-container"  *ngIf="course.words.length > 0">
+    <select class="table-control-select" (change)="changeSelectedTableAction($event)">
+        <option value="null">(Select Action)</option>
+        <option value="suspend">Suspend Word(s)</option>
+        <option value="restore">Restore Word(s)</option>
+        <option value="edit">Edit Word</option>
+        <option value="reset">Reset Word(s)</option>
+    </select>
+    <button type="button" class="table-control-submit" (click)="executeTableAction()">Submit</button>
+</div>
+<div class="mat-elevation-z8 word-table" *ngIf="course.words.length > 0;">
+    <table mat-table [dataSource]="dataSource">
+
+        <!-- Japanese Vocab Column -->
+        <ng-container matColumnDef="Japanese">
+            <th mat-header-cell *matHeaderCellDef> Japanese Vocab </th>
+            <td mat-cell *matCellDef="let word">
+                <input class="word-checkbox" type="checkbox" (change)="toggleWordStatus(word)">
+                {{word.japaneseVocab}}
+                [{{word.kanaVocab}}]
+            </td>
+        </ng-container>
+    
+        <!-- English Vocab Column -->
+        <ng-container matColumnDef="English">
+            <th mat-header-cell *matHeaderCellDef> English </th>
+            <td mat-cell *matCellDef="let word"> {{word.englishVocab}} </td>
+        </ng-container>
+    
+        <!-- Sentence Column -->
+        <ng-container matColumnDef="Sentence">
+            <th mat-header-cell *matHeaderCellDef> Sentence </th>
+            <td mat-cell *matCellDef="let word">
+                <p *ngFor="let sentence of word.sentences">
+                    <span
+                    *ngIf="sentence.sentenceText != null &&
+                    sentence.sentenceType.type != 'Furigana'">{{sentence.sentenceText}}</span>
+                </p>
+            </td>
+        </ng-container>
+    
+        <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+        <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
+    </table>
+
+    <mat-paginator (page)="checkForAdditionalTableData($event)" [pageSizeOptions]="[5, 10, 25]" showFirstLastButtons></mat-paginator>
+</div>
